It definitely looks like Octree was a waste.
I must be using it incorrectly in some way, because if it can only do what it appears to do as documented, it's essentially worthless... Right now it can tell me what Movers I have added to the Octree, and when I query with a non-overlapping Mover, it just returns... all the Movers in the Octree. Regardless of whether they touch the queried area. Wat...